Complete fixture symbol background generation cleanup
